fix: adding products to cart
