+++ src/features/login-form/index.test.tsx
@@ -12,7 +12,6 @@ import userEvent from '@testing-library/user-event';
 import LoginForm from './index';
 import { UserContextProvider } from '../../components/contexts/user/context-provider';
 import { LoadingContextProvider } from '../../components/contexts/loading/context-provider';
-import { CartContextProvider } from '../../components/contexts/cart/context-provider';
 
 describe('Ensure that input validation checks are performed in real-time when the user enters their information.', () => {
   let emailField: HTMLInputElement;
@@ -36,15 +35,13 @@ describe('Ensure that input validation checks are performed in real-time when th
       getParentRoute: () => rootRoute,
       path: '/',
       component: () => (
-        <CartContextProvider>
-          <UserContextProvider>
-            <LoadingContextProvider>
-              <div data-testid="is-rendered">
-                <LoginForm />
-              </div>
-            </LoadingContextProvider>
-          </UserContextProvider>
-        </CartContextProvider>
+        <UserContextProvider>
+          <LoadingContextProvider>
+            <div data-testid="is-rendered">
+              <LoginForm />
+            </div>
+          </LoadingContextProvider>
+        </UserContextProvider>
       ),
     });
     const testRouter = createRouter({
